DUC streamer API improvement (#67)
#### Refactor the DUC streamer implementation

##### Motivation

Passing a websocket address to the streamer will get you a V3 streamer;
`tcp` or `tis` addresses will get a V1 or V2 streamer depending on the
server. At that point, the API contract is not clear:

- Only V1 or V2 streamers want keep alives
- Only V1 or V2 streamers support `SendPayload(...)`
- Only V3 streamers support `SendRawData(...)`
- It's up to the consumer to check `IsRawData` and call the appropriate
function
- Working in the code is somewhat complicated, as the `DucStreamer`
class contains functions and methods for all three versions, and the
control flow is managed with conditionals
- Frustrated that I wasted time fixing bugs in `SendPayload` when
actually I should've been calling `SendRawData`. None of this is obvious
without an understanding of the class, which is difficult because of the
above
- The `DucStreamer` class is hard to extend and modify because of the
above
- There is a lot of low-hanging optimization possible here that's much
easier with an improved structure

##### Implementation

Add a base `BaseDucStreamerBackend` and then concrete implementations
for V1, V2 and V3 streamers.

`DucStreamer` will connect and create the appropriate backend
accordingly. Because the difference between V1 and V2 is only known at
connection time, this is handled with a `DucStreamerConnector`. Function
calls are forwarded to the underlying implementation. `SendPayload` and
`SendRawData` are mapped between the streamers, although in certain
cases it's unavoidable to copy the data or start a new task so this
should be avoided. I added a `SendRawDataAsync` function.

In terms of starting to modernize the API, I added the `netstandard2.1`
target and when used we get a pair of new common functions:

```
Task WriteAsync(byte[] buffer, int offset, int count, CancellationToken cancellationToken = default) ;
Task WriteAsync(IntPtr buffer, int length, CancellationToken cancellationToken = default);
```

These are properly async, zero-copy and will use the according
underlying function in the streamer.

The V1/V2 implementation of `WriteAsync` fixes a bug in the existing
`SendPayload` - we call `Stream.GetBuffer()` and then operate on the
length of the returned byte array. In fact, the byte array backing the
buffer is almost certainly over allocated and we write redundant zeroes
to the payload. The `WriteAsync` function properly operates on the
length of the stream. The existing `SendPayload` functions are not fixed
- I added a comment, but I'm worried about backward compatibility. The
happy path moving forward is streamer V3 + `WriteAsync`.

There were public fields in `DucStreamer` that I marked as obsolete and
replaced with properties. Otherwise the public API is unchanged for
backward compatibility. Moving forward, this will make it easier to
extend and improve the streamer.